requirements.txt
comtypes==1.4.10
easyocr==1.7.2
filelock==3.18.0
fsspec==2025.3.0
imageio==2.37.0
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
ninja==1.11.1.4
numpy==2.2.4
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
pyclipper==1.3.0.post6
pypiwin32==223
pyspellchecker==0.8.2
python-bidi==0.6.6
pyttsx3==2.98
pywin32==310
PyYAML==6.0.2
scikit-image==0.25.2
scipy==1.15.2
setuptools==77.0.3
shapely==2.0.7
sympy==1.13.1
tifffile==2025.3.13
torch==2.6.0
torchvision==0.21.0
typing_extensions==4.12.2